♻️Refactor: interfaces layer package 구조 변경  - 멘토링 피드백 반영   . 다른 layer 계층과 패키지 일관성 통일 -> 가독성 향상
