--- src/main/java/kr/hhplus/be/interfaces/dto/order/OrderItemRequest.java
@@ -1,15 +1,0 @@
-package kr.hhplus.be.interfaces.dto.order;
-
-import kr.hhplus.be.application.order.dto.OrderItemInfo;
-import lombok.Getter;
-
-@Getter
-public class OrderItemRequest {
-    private Long productId;
-    private Integer quantity;
-    private Integer price;
-
-    public OrderItemInfo toOrderItemInfo() {
-        return new OrderItemInfo(productId, quantity, price);
-    }
-}
