--- src/main/java/kr/hhplus/be/interfaces/dto/product/ProductSelectRequest.java
@@ -1,31 +1,0 @@
-package kr.hhplus.be.interfaces.dto.product;
-
-import jakarta.validation.constraints.NotNull;
-import kr.hhplus.be.domain.product.dto.ProductSearchDTO;
-import kr.hhplus.be.domain.product.enumtype.Category;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProductSelectRequest {
-    private String productName;
-    private int startPrice;
-    private int endPrice;
-    private Category category;
-
-    // paging info
-    @NotNull
-    private int page;
-    @NotNull
-    private int size;
-
-    public ProductSearchDTO toDTO() {
-        return ProductSearchDTO.builder()
-                .productName(this.productName)
-                .startPrice(this.startPrice)
-                .endPrice(this.endPrice)
-                .category(this.category).build();
-    }
-}
