--- src/main/java/kr/hhplus/be/interfaces/dto/user/ChargeResponse.java
@@ -1,11 +1,0 @@
-package kr.hhplus.be.interfaces.dto.user;
-
-import lombok.*;
-
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class ChargeResponse {
-    private int balance;
-}
